Remove mock's default init when protocol already requires an empty init
