+++ Sources/MockedClient/VariadicParameters.swift
@@ -1,0 +1,19 @@
+//
+//  VariadicParameters.swift
+//  MockedClient
+//
+//  Created by Gray Campbell on 3/30/25.
+//
+
+import Foundation
+public import Mocked
+
+/// A protocol for verifying Mocked's handling of variadic parameters.
+///
+/// - Important: Please only use this protocol for permanent verification of
+///   Mocked's handling of variadic parameters. For temporary testing of
+///   Mocked's expansion, use the `Playground` protocol in `main.swift`.
+@Mocked
+public protocol VariadicParameters {
+    func method(strings: String..., integers: Int...)
+}
